code refactor - added seperate folder for services and moved appointmentBookingService and clinicportal calls to new folder
